Now allows building/running of python images/code using -p option on spark 3.5 or better (#34)
Modified createImage and submitSparkPy scripts to take a -p option.

This option causes the createImage script to build an image that
includes the python executables. I made this an option, instead of
including python in every build, because it causes the image build
process to go from 6 seconds to 45.

It causes the submitSparkPy script to change the submitted parameters so
they support running the spark pi.py example.

---------

Signed-off-by: George Jahad <github@blackbirdsystems.net>