Developer e2e test script (#39)
This is a basic script for developer testing convenience on a local
Armada setup. It's intended for developer interactive use, and not for
automated/CI checks. Some parts of the verification logic have been
appropriated from `.github/actions/e2e/action.yaml.

It utilizes the settings found in `scripts/config.sh`, plus some
additional settings at the beginning of this script. It checks for the
existence of the Armada cluster, the test queue, and the local
spark-armada image (in Docker). If Armada is not available, it attempts
to start it using the `armada-operator`; if the queue is not found, it
creates it. If the spark-armada image is not found, the script exits
(it's assumed the user would prefer to build it separately, outside this
script). It does the usual verification of the Spark Driver and
Executor(s) after submission.

Fixes https://github.com/G-Research/spark/issues/29

---------

Signed-off-by: Rich Scott <richscott@sent.com>