Fix a few jobtemplate issues (#63)
This PR addresses issues that I've found with the JobTemplate code.
Dejan and I discussed most of them in the JobTemplate PR.

The main one is that the mergeDriverTemplate/newDriverJobItem() and
mergeExecutorTemplate/newExecutorJobItem() methods had a lot of
duplicate code and were getting out of sync. So I refactored them to
remove newDriverJobItem()/newExecutorJobItem()

I also made runAsUser overrideable with a template.
The rest of the changes are pretty minor, fixing up magic numbers, log
messages and a few other little things I noticed when running the
internal tests.

---------

Signed-off-by: George Jahad <github@blackbirdsystems.net>
Co-authored-by: Enrico Minack <github@enrico.minack.dev>