Add feature step support (#65)
## Intro

Spark on K8S uses configuration blocks called "Feature Steps" They are
written with the fabric8 library, which isn't supported by Armada.

This PR allows their use by invoking them with the fabric8 api, and then
serializing them to yaml, and then unmarshalling that into the api we
can use with Armada.

The heart of the code is here:

https://github.com/GeorgeJahad/armada-spark/blob/43efb3cbb20ce6ed9a7988085281373a3123de61/src/main/scala/org/apache/spark/deploy/armada/submit/ArmadaClientApplication.scala#L247-L351

## Other changes

1. Added a custom deserializer for the Quantity type in job templates.

2. Removed the use of scala.util.Using as it doesn't exist in scala 2.12


## Issues
I've run into some issues with the volume mount unmarshalling; job
templates seem to have the same issue:
https://github.com/G-Research/spark/issues/109

We will have to look further into that.

---------

Signed-off-by: George Jahad <github@blackbirdsystems.net>
Co-authored-by: Enrico Minack <github@enrico.minack.dev>