fix: issue where certain podspec fields weren't getting parsed by the JobTemplateLoader (#72)
## Problem

The Armada Scala client is generated using ScalaPB, which generates the
`PodSpec` field in the `JobSubmitRequestItem` from the k8s proto files,
and they differ a bit from the standard PodSpec manifest and from the
`fabric` PodSpec class.

Some fields are wrapped (like volume types) and fail to get
deserialized.

## Solution

This PR edits the `JobTemplateLoader` to register a custom deserializer
which uses the `PodSpecConverted` to map from the proto-generated
classes to the fabric8 classes and vice-versa.

## Additional Notes

`JobTemplateLoader` had some unneccesary compexities, this PR also
cleans up the implementation and the tests a bit.

Closes
https://github.com/orgs/G-Research/projects/36/views/1?pane=issue&itemId=119248104&issue=G-Research%7Cspark%7C109

Signed-off-by: Dejan Zele Pejchev <pejcev.dejan@gmail.com>