dynamic allocation implementation (#82)
## Summary
Dynamic Allocation is implemented by supporting the callback functions,
that spark core calls, to add/delete executors. Those functions are
implemented here: ArmadaClusterManagerBackend. The most important ones
are: doRequestTotalExecutors() which sets the total number of executors
that should be running, and doKillExecutors()

The other important part of this PR is the ArmadaEventWatcher, which
watches the armada event stream for this job set and reports on the
starting and stopping of executors from the Armada side.

addresses: https://github.com/G-Research/spark/issues/137 and
https://github.com/G-Research/spark/issues/133

---------

Signed-off-by: George Jahad <github@blackbirdsystems.net>
Co-authored-by: Enrico Minack <github@enrico.minack.dev>